requirements.txt
openai
python-dotenv
requests
filelock
colorama
pathlib